requirement.txt
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
voluptuous==0.14.1
wcwidth==0.2.8
websocket-client==1.7.0
Werkzeug==3.0.0
wrapt==1.14.1
absl-py==0.13.0
asttokens==2.4.0
astunparse==1.6.3
awscli==1.31.12
backcall==0.2.0
boto3==1.18.12
botocore==1.21.12
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.0
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.16
easydict==1.10
efficientnet-pytorch==0.7.1
einops==0.7.0
executing==2.0.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
imageio==2.31.5
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
jaraco.classes==3.3.0
jax==0.4.23
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.12.0
Keras-Preprocessing==1.1.2
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
ml-dtypes==0.2.0
more-itertools==10.1.0
mpmath==1.3.0
natsort==8.4.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.8.2
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.14.1
torch==2.0.1
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.11.2
tensorflow==2.12.0
transformers==4.34.1
yolo==0.3.2